fix(bungee)!: check protocol state while trying to log in
This fixes the bug related to a connection reset if the backend gives a response too soon and the user is not on the game protocol state (login or config phase).

You must need to add the jar locally to the bungeecord module in the /libs/ to be able to build the jar!